switch documentation job to new PTI
This is a mechanically generated patch to switch the documentation
jobs to use the new PTI versions of the jobs as part of the
python3-first goal.

See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

Update tox.ini so that tox -e docs works. Remove function exectution
from documentation - the function section includes it, this caused a
duplicate section.

Story: #2002586
Task: #24326

Co-Authored-By: Andreas Jaeger <jaegerandi@gmail.com>
Co-Authored-By: Vieri <15050873171@163.com>
Change-Id: I8604aba14b4e9c35e446e53d997f63d12a65e199