Fix broken function version during webhook update
When updating a webhook which uses a function_version other than 0
using qinlingclient, the function_version of that webhook would be
reset to 0.

This is caused by the argument 'function_version' of the webhook
update command is set to a default value 0, so it is always passed
to the Qinling API even though user doesn't specify it.

Change-Id: Ic1071a5ba7bfd12161a423c99d888857a786a1ba
Story: 2003478
Task: 24744