Retire the python-qinlingclient project
As announced in openstack-discuss ML[1], Qinling project
is retiring in Wallaby cycle.

This commit retires the python-qinlingclient repository as per
process deinfed in project-guide[2]. Anyone would like to
maintain it again, please revert back this commit and propose
the re-adding it to governance.

The community wishes to express our thanks and appreciation to all of
those who have contributed to the python-qinlingclient project over
the years.

Depends-On: https://review.opendev.org/c/openstack/project-config/+/764520
Needed-By: https://review.opendev.org/c/openstack/governance/+/764523
Change-Id: Idc709077f3f1fbbc81795b32a0cf5fc005f3dd27